Moved resource link to sync command
