+++ src/Console/SyncCommand.php
@@ -29,6 +29,7 @@ class SyncCommand extends Command
     public function handle()
     {
         try {
+            dispatch(new \Fusion\Console\Actions\SyncResources);
             dispatch(new \Fusion\Console\Actions\SyncModules);
             dispatch(new \Fusion\Console\Actions\SyncExtensions);
             dispatch(new \Fusion\Console\Actions\SyncSettings);
